+++ example/dapp/lib/pages/connect_page.dart
@@ -1,0 +1,249 @@
+import 'package:fl_toast/fl_toast.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+import 'package:qr_flutter/qr_flutter.dart';
+import 'package:walletconnect_flutter_v2/walletconnect_flutter_v2.dart';
+import 'package:walletconnect_flutter_v2_dapp/models/chain_metadata.dart';
+import 'package:walletconnect_flutter_v2_dapp/utils/constants.dart';
+import 'package:walletconnect_flutter_v2_dapp/utils/crypto/chain_data.dart';
+import 'package:walletconnect_flutter_v2_dapp/utils/crypto/helpers.dart';
+import 'package:walletconnect_flutter_v2_dapp/utils/string_constants.dart';
+import 'package:walletconnect_flutter_v2_dapp/widgets/chain_button.dart';
+
+class ConnectPage extends StatefulWidget {
+  const ConnectPage({
+    super.key,
+    required this.web3App,
+  });
+
+  final Web3App web3App;
+
+  @override
+  ConnectPageState createState() => ConnectPageState();
+}
+
+class ConnectPageState extends State<ConnectPage> {
+  bool _testnetOnly = false;
+  final List<ChainMetadata> _selectedChains = [];
+
+  void setTestnet(bool value) {
+    if (value != _testnetOnly) {
+      _selectedChains.clear();
+    }
+    _testnetOnly = value;
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    // Build the list of chain buttons, clear if the textnet changed
+    final List<ChainMetadata> chains =
+        _testnetOnly ? ChainData.testChains : ChainData.mainChains;
+
+    List<Widget> chainButtons = [];
+
+    for (final ChainMetadata chain in chains) {
+      // Build the button
+      chainButtons.add(
+        ChainButton(
+          chain: chain,
+          onPressed: () {
+            setState(() {
+              if (_selectedChains.contains(chain)) {
+                _selectedChains.remove(chain);
+              } else {
+                _selectedChains.add(chain);
+              }
+            });
+          },
+          selected: _selectedChains.contains(chain),
+        ),
+      );
+    }
+
+    // Add a connect button
+    chainButtons.add(
+      Container(
+        width: double.infinity,
+        height: StyleConstants.linear48,
+        margin: const EdgeInsets.symmetric(
+          vertical: StyleConstants.linear8,
+        ),
+        child: ElevatedButton(
+          onPressed: () => _onConnect(_selectedChains),
+          style: ButtonStyle(
+            backgroundColor: MaterialStateProperty.all<Color>(
+              StyleConstants.primaryColor,
+            ),
+            shape: MaterialStateProperty.all<RoundedRectangleBorder>(
+              RoundedRectangleBorder(
+                borderRadius: BorderRadius.circular(
+                  StyleConstants.linear8,
+                ),
+              ),
+            ),
+          ),
+          child: const Text(
+            StringConstants.connect,
+            style: StyleConstants.buttonText,
+          ),
+        ),
+      ),
+    );
+
+    return Center(
+      child: Container(
+        // color: StyleConstants.primaryColor,
+        padding: const EdgeInsets.all(
+          StyleConstants.linear8,
+        ),
+        constraints: const BoxConstraints(
+          maxWidth: StyleConstants.maxWidth,
+        ),
+        child: ListView(
+          children: <Widget>[
+            const SizedBox(
+              height: StyleConstants.linear48,
+            ),
+            const Text(
+              StringConstants.appTitle,
+              style: StyleConstants.titleText,
+              textAlign: TextAlign.center,
+            ),
+            const SizedBox(
+              height: StyleConstants.linear48,
+            ),
+            const Text(
+              StringConstants.selectChains,
+              style: StyleConstants.subtitleText,
+              textAlign: TextAlign.center,
+            ),
+            const SizedBox(
+              height: StyleConstants.linear24,
+            ),
+            SizedBox(
+              height: StyleConstants.linear48,
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  const Text(
+                    StringConstants.testnetsOnly,
+                    style: StyleConstants.buttonText,
+                  ),
+                  Switch(
+                    value: _testnetOnly,
+                    onChanged: (value) {
+                      setState(() {
+                        _testnetOnly = value;
+                      });
+                    },
+                  ),
+                ],
+              ),
+            ),
+            Column(
+              mainAxisAlignment: MainAxisAlignment.center,
+              crossAxisAlignment: CrossAxisAlignment.center,
+              children: chainButtons,
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+
+  void _onConnect(
+    List<ChainMetadata> chains,
+  ) async {
+    // Use the chain metadata to build the required namespaces:
+    // Get the methods, get the events
+    final Map<String, RequiredNamespace> requiredNamespaces = {};
+    final Map<String, dynamic> data = {};
+    for (final chain in chains) {
+      final RequiredNamespace rNamespace = RequiredNamespace(
+        chains: [chain.chainId],
+        methods: getChainMethods(chain.type),
+        events: getChainEvents(chain.type),
+      );
+      requiredNamespaces[chain.chainId.split(':')[0]] = rNamespace;
+      data[chain.chainId] = rNamespace.toJson();
+    }
+    debugPrint(data.toString());
+
+    // Send off a connect
+    final ConnectResponse res = await widget.web3App.connect(
+      requiredNamespaces: requiredNamespaces,
+    );
+    // print(res.uri!.toString());
+    await _showQrCode(res);
+  }
+
+  Future<void> _showQrCode(
+    ConnectResponse response,
+  ) async {
+    // Show the QR code
+    showDialog(
+      context: context,
+      builder: (context) {
+        return AlertDialog(
+          title: const Text(
+            StringConstants.scanQrCode,
+            style: StyleConstants.titleText,
+            textAlign: TextAlign.center,
+          ),
+          content: SizedBox(
+            width: 300,
+            height: 350,
+            child: Center(
+              child: Column(
+                children: [
+                  QrImage(
+                    data: response.uri!.toString(),
+                  ),
+                  const SizedBox(
+                    height: StyleConstants.linear16,
+                  ),
+                  ElevatedButton(
+                    onPressed: () async {
+                      await Clipboard.setData(
+                        ClipboardData(
+                          text: response.uri!.toString(),
+                        ),
+                      );
+                      await showPlatformToast(
+                        child: const Text(
+                          StringConstants.urlCopiedToClipboard,
+                        ),
+                        context: context,
+                      );
+                    },
+                    child: const Text(
+                      'Copy URL to Clipboard',
+                    ),
+                  ),
+                ],
+              ),
+            ),
+          ),
+        );
+      },
+    );
+
+    try {
+      final SessionData _ = await response.session.future;
+      await showPlatformToast(
+        child: const Text(
+          StringConstants.connectionEstablished,
+        ),
+        context: context,
+      );
+    } catch (e) {
+      debugPrint(e.toString());
+      await showPlatformToast(
+        child: const Text(
+          StringConstants.connectionEstablished,
+        ),
+        context: context,
+      );
+    }
+  }
+}