+++ example/dapp/lib/utils/constants.dart
@@ -1,0 +1,49 @@
+import 'package:flutter/material.dart';
+
+class Constants {
+  static const projectId = '7e984f90b95f0236d3c12d791537f233';
+
+  static const smallScreen = 640;
+}
+
+class StyleConstants {
+  static const Color primaryColor = Color.fromARGB(255, 16, 165, 206);
+  static const Color secondaryColor = Color(0xFF1A1A1A);
+
+  // Linear
+  static const double linear8 = 8;
+  static const double linear16 = 16;
+  static const double linear24 = 24;
+  static const double linear32 = 32;
+  static const double linear48 = 48;
+  static const double linear56 = 56;
+  static const double linear72 = 72;
+  static const double linear80 = 80;
+
+  // Magic Number
+  static const double magic10 = 10;
+  static const double magic14 = 14;
+  static const double magic20 = 20;
+  static const double magic40 = 40;
+  static const double magic64 = 64;
+
+  // Width
+  static const double maxWidth = 400;
+
+  // Text styles
+  static const TextStyle titleText = TextStyle(
+    color: Colors.black,
+    fontSize: magic40,
+    fontWeight: FontWeight.w600,
+  );
+  static const TextStyle subtitleText = TextStyle(
+    color: Colors.black,
+    fontSize: linear24,
+    fontWeight: FontWeight.w600,
+  );
+  static const TextStyle buttonText = TextStyle(
+    color: Colors.black,
+    fontSize: magic14,
+    fontWeight: FontWeight.w600,
+  );
+}