+++ example/dapp/lib/utils/crypto/eip155_data.dart
@@ -1,0 +1,59 @@
+enum EIP155Methods {
+  personalSign,
+  ethSign,
+  ethSignTransaction,
+  ethSignTypedData,
+  ethSignTypedDataV3,
+  ethSignTypedDataV4,
+  ethSendRawTransaction,
+  ethSendTransaction,
+}
+
+enum EIP155Events {
+  chainChanged,
+  accountsChanged,
+}
+
+extension EIP155MethodsX on EIP155Methods {
+  String? get value => EIP155Data.methods[this];
+}
+
+extension EIP155MethodsStringX on String {
+  EIP155Methods? toEip155Method() {
+    final entries = EIP155Data.methods.entries.where(
+      (element) => element.value == this,
+    );
+    return (entries.isNotEmpty) ? entries.first.key : null;
+  }
+}
+
+extension EIP155EventsX on EIP155Events {
+  String? get value => EIP155Data.events[this];
+}
+
+extension EIP155EventsStringX on String {
+  EIP155Events? toEip155Event() {
+    final entries = EIP155Data.events.entries.where(
+      (element) => element.value == this,
+    );
+    return (entries.isNotEmpty) ? entries.first.key : null;
+  }
+}
+
+class EIP155Data {
+  static final Map<EIP155Methods, String> methods = {
+    EIP155Methods.personalSign: 'personal_sign',
+    EIP155Methods.ethSign: 'eth_sign',
+    EIP155Methods.ethSignTransaction: 'eth_signTransaction',
+    EIP155Methods.ethSignTypedData: 'eth_signTypedData',
+    EIP155Methods.ethSignTypedDataV3: 'eth_signTypedData_v3',
+    EIP155Methods.ethSignTypedDataV4: 'eth_signTypedData_v4',
+    EIP155Methods.ethSendRawTransaction: 'eth_sendRawTransaction',
+    EIP155Methods.ethSendTransaction: 'eth_sendTransaction',
+  };
+
+  static final Map<EIP155Events, String> events = {
+    EIP155Events.chainChanged: 'personal_sign',
+    EIP155Events.accountsChanged: 'eth_sign',
+  };
+}