+++ example/dapp/lib/utils/crypto/solana_data.dart
@@ -1,0 +1,17 @@
+enum SolanaMethods {
+  solanaSignTransaction,
+  solanaSignMessage,
+}
+
+enum SolanaEvents {
+  none,
+}
+
+class SolanaData {
+  static final Map<SolanaMethods, String> methods = {
+    SolanaMethods.solanaSignTransaction: 'solana_signTransaction',
+    SolanaMethods.solanaSignMessage: 'solana_signMessage'
+  };
+
+  static final Map<SolanaEvents, String> events = {};
+}