+++ lib/apis/auth_api/auth_client.dart
@@ -5,6 +5,7 @@ import 'package:walletconnect_flutter_v2/apis/auth_api/auth_engine.dart';
 import 'package:walletconnect_flutter_v2/apis/auth_api/i_auth_client.dart';
 import 'package:walletconnect_flutter_v2/apis/auth_api/i_auth_engine.dart';
 import 'package:walletconnect_flutter_v2/apis/auth_api/models/auth_client_events.dart';
+import 'package:walletconnect_flutter_v2/apis/core/core.dart';
 import 'package:walletconnect_flutter_v2/apis/core/store/generic_store.dart';
 import 'package:walletconnect_flutter_v2/apis/core/store/i_generic_store.dart';
 import 'package:walletconnect_flutter_v2/apis/auth_api/models/auth_client_models.dart';
@@ -12,6 +13,7 @@ import 'package:walletconnect_flutter_v2/apis/auth_api/utils/auth_constants.dart
 import 'package:walletconnect_flutter_v2/apis/core/i_core.dart';
 import 'package:walletconnect_flutter_v2/apis/models/basic_models.dart';
 import 'package:walletconnect_flutter_v2/apis/core/pairing/utils/pairing_models.dart';
+import 'package:walletconnect_flutter_v2/apis/utils/constants.dart';
 
 class AuthClient implements IAuthClient {
   bool _initialized = false;
@@ -44,14 +46,19 @@ class AuthClient implements IAuthClient {
   late IAuthEngine engine;
 
   static Future<AuthClient> createInstance({
-    required ICore core,
+    required String projectId,
+    String relayUrl = WalletConnectConstants.DEFAULT_RELAY_URL,
     required PairingMetadata metadata,
+    bool memoryStore = false,
   }) async {
     final client = AuthClient(
-      core: core,
+      core: Core(
+        projectId: projectId,
+        relayUrl: relayUrl,
+        memoryStore: memoryStore,
+      ),
       metadata: metadata,
     );
-
     await client.init();
 
     return client;